backend/requirements.txt
fastapi
uvicorn
redis
yfinance
pandas
ta
pyautogen
python-dotenv
requests
fpdf2
